requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
duckduckgo-search>=7.0.0
colorama>=0.4.6
